Add support for basic shapes (circle, quad)
All shapes are an extension of the base `Shape` class, which describes a
shape as a collection of vertices. The vertices are two-dimensional arrays
of floating point numbers (vectors) as defined in `Vec2`.

The basic shapes are a circle (`Circle`) and rectangle (`Quad`). These
provide a few extra convenience accessors that make it easier to use the
shapes for collision / intersection detection.

Narrow-phase collision detection algorithims define the most generic
collision algorithms for the described base shapes, namely circle-circle
collision (based on the radius of a circle) and quad-quad collision (based
on the minimum and maximum x, y values).